64-bit fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21020 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/verbosity.cpp
@@ -60,7 +60,7 @@ csVerbosityParser::csVerbosityParser (const char* s)
 bool csVerbosityParser::Error(char const* msg, char const* s, size_t pos)
 {
   csPrintfErr("ERROR: Verbosity parser: %s: %-*.*s<<ERROR>>%s\n",
-	      msg, pos, pos, s, s + pos);
+	      msg, (int)pos, (int)pos, s, s + pos);
   return false;
 }
